requirements.txt
scikit-surprise
pandas
numpy
scikit-learn
re
unidecode
pickle
sentence-transformers